collector/lmstat: consolidate linux and windows implementations
