+++ collector/lmstat_test.go
